Refactor extract With command from CLI
