+++ cmd/main.go
@@ -37,15 +37,8 @@ func main() {
 		_ = pair.Version(os.Stdout, os.Stderr)
 		os.Exit(0)
 	case "with":
-		err := configureGit(template)
-		if err != nil {
-			return
-		}
-		pair := os.Args[2]
-		overwrite(template, fmt.Sprintf("\n\nCo-Authored-By: %s\n", pair))
-		if err != nil {
-			return
-		}
+		_ = pair.With(os.Stdout, os.Stderr, os.Args[2])
+		os.Exit(0)
 	case "stop":
 		_ = remove(template)
 		err := unconfigureGit()
@@ -67,34 +60,6 @@ func checkArgs(args []string) error {
 		return nil
 	}
 	return &argumentsError{}
-}
-
-func configureGit(template string) (err error) {
-	err = unconfigureGit()
-	if err != nil {
-		return
-	}
-	cmd := exec.Command("git", "config", "--global", "--add", "commit.template", template)
-	err = cmd.Run()
-	if err != nil {
-		return
-	}
-	return
-}
-
-func overwrite(template, pair string) (err error) {
-	_ = remove(template)
-	f, err := os.OpenFile(template, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
-	if err != nil {
-		return
-	}
-	if _, err = f.Write([]byte(pair)); err != nil {
-		return
-	}
-	if err = f.Close(); err != nil {
-		return
-	}
-	return
 }
 
 func remove(template string) (err error) {
